Enable building containers from tags or branches
Currently building the container will always build from the tip of master
on every repo.  The branch or tag ref needs to be passed to align with
other manageiq repos that are required for components or features.
To build with changes in your local source tree, uncomment the ADD and
comment the RUN for adding the manageiq source

[skip ci]


(transferred from ManageIQ/manageiq@a96ac42ef961a97fc578dd3fbc821680d32cb131)